+++ mobile/src/terminal/terminal-webview-html.ts
@@ -1,28 +1,34 @@
-import { TERMINAL_HTML_FRAGMENT_01 } from './terminal-webview-html/fragment-01'
-import { TERMINAL_HTML_FRAGMENT_02 } from './terminal-webview-html/fragment-02'
-import { TERMINAL_HTML_FRAGMENT_03 } from './terminal-webview-html/fragment-03'
-import { TERMINAL_HTML_FRAGMENT_04 } from './terminal-webview-html/fragment-04'
-import { TERMINAL_HTML_FRAGMENT_05 } from './terminal-webview-html/fragment-05'
-import { TERMINAL_HTML_FRAGMENT_06 } from './terminal-webview-html/fragment-06'
-import { TERMINAL_HTML_FRAGMENT_07 } from './terminal-webview-html/fragment-07'
-import { TERMINAL_HTML_FRAGMENT_08 } from './terminal-webview-html/fragment-08'
-import { TERMINAL_HTML_FRAGMENT_09 } from './terminal-webview-html/fragment-09'
-import { TERMINAL_HTML_FRAGMENT_10 } from './terminal-webview-html/fragment-10'
+import { TERMINAL_HTML_DOCUMENT_SHELL } from './terminal-webview-html/document-shell'
+import { TERMINAL_HTML_RUNTIME_STATE_AND_TEXT_SCALING } from './terminal-webview-html/runtime-state-and-text-scaling'
+import { TERMINAL_HTML_FIT_SCALE_AND_WRITE_QUEUE } from './terminal-webview-html/fit-scale-and-write-queue'
+import { TERMINAL_HTML_INIT_AND_WRITE } from './terminal-webview-html/terminal-init-and-write'
+import { TERMINAL_HTML_HOST_MESSAGE_ROUTER } from './terminal-webview-html/host-message-router'
+import { TERMINAL_HTML_SELECTION_STATE_AND_EVICTION } from './terminal-webview-html/selection-state-and-eviction'
+import { TERMINAL_HTML_OBSERVERS_AND_MODE_MIRRORING } from './terminal-webview-html/term-observers-and-mode-mirroring'
+import { TERMINAL_HTML_MOUSE_REPORT_AND_SCROLL_ROUTING } from './terminal-webview-html/mouse-report-and-scroll-routing'
+import { TERMINAL_HTML_SMOOTH_SCROLL_AND_CELL_GEOMETRY } from './terminal-webview-html/smooth-scroll-and-cell-geometry'
+import { TERMINAL_HTML_SELECTION_OVERLAY } from './terminal-webview-html/selection-overlay'
+import { TERMINAL_HTML_SURFACE_TOUCH_GESTURES } from './terminal-webview-html/surface-touch-gestures'
+import { TERMINAL_HTML_MESSAGE_BRIDGE_AND_DOCUMENT_CLOSE } from './terminal-webview-html/message-bridge-and-document-close'
 
 export { MOBILE_TERMINAL_CARET_OPTIONS } from './terminal-webview-html/theme'
 
-// Why: keep the document source stable while each script/style concern remains independently reviewable.
+// Why: keep the document source stable while each script/style concern remains independently
+// reviewable. Boundaries can only fall where the emitted document allows, so a few modules
+// carry a second concern noted at the top of the file.
 export const XTERM_HTML = [
-  TERMINAL_HTML_FRAGMENT_01,
-  TERMINAL_HTML_FRAGMENT_02,
-  TERMINAL_HTML_FRAGMENT_03,
-  TERMINAL_HTML_FRAGMENT_04,
-  TERMINAL_HTML_FRAGMENT_05,
-  TERMINAL_HTML_FRAGMENT_06,
-  TERMINAL_HTML_FRAGMENT_07,
-  TERMINAL_HTML_FRAGMENT_08,
-  TERMINAL_HTML_FRAGMENT_09,
-  TERMINAL_HTML_FRAGMENT_10
+  TERMINAL_HTML_DOCUMENT_SHELL,
+  TERMINAL_HTML_RUNTIME_STATE_AND_TEXT_SCALING,
+  TERMINAL_HTML_FIT_SCALE_AND_WRITE_QUEUE,
+  TERMINAL_HTML_INIT_AND_WRITE,
+  TERMINAL_HTML_HOST_MESSAGE_ROUTER,
+  TERMINAL_HTML_SELECTION_STATE_AND_EVICTION,
+  TERMINAL_HTML_OBSERVERS_AND_MODE_MIRRORING,
+  TERMINAL_HTML_MOUSE_REPORT_AND_SCROLL_ROUTING,
+  TERMINAL_HTML_SMOOTH_SCROLL_AND_CELL_GEOMETRY,
+  TERMINAL_HTML_SELECTION_OVERLAY,
+  TERMINAL_HTML_SURFACE_TOUCH_GESTURES,
+  TERMINAL_HTML_MESSAGE_BRIDGE_AND_DOCUMENT_CLOSE
 ].join('')
 
 export const XTERM_WEBVIEW_SOURCE = { html: XTERM_HTML }
